Improvements to settings
contentMap allows you to supply a search filter for each page, so that one page
can map to multiple pieces of content. E.g. a news article listing page.

Excludes is also now a list, so you can exclude multiple urls.

The image download handling prevents downloading images even when they are the target of a link (not just an image).