Make the crawling accessible via the web page.  (#1)
* Update build system to compile the web javascript.

Create a single bundle.js file that can be included from the webpage,
and contains all the node dependencies, built with browserify.

Include Mocha for unit tests.

* New split build source files.

Storage contains IndexDB and local file storage for the web and cli versions

crawler.js is the main class and the "web" and "cli" wrappers handle loading via node or in a web page.

* Add comments to the crawler.

* Add Mocha tests for the crawler.

* Improve handling of simplification.

This aims to remove redundant parent nodes.

* Detect duplicate content across pages.

Remove the duplicates - this implies things like headers, footers or navigation that are repeated across pages.

* Split the log events

Use an event to communicate from the crawler to the web page.

* Clean more common navigation elements.

* Add a setting to map urls to content types.

* Use the public CORS proxy.

* Include one "example" auth key.

* Move crawler settings to it's own class.

* Move urlFilter function into settings.

* Move complete handler to own class method.

* Move duplicate stripping to it's own class.

* Add unit test for duplicate removal.

* Add all supported options to cli

* Add a php script to allow updating the server

This lets you remove server sites, and move sites from local storage to the server.

* Add client management to the web interface.

* Bring back the form handling.

* Improvements to settings

contentMap allows you to supply a search filter for each page, so that one page
can map to multiple pieces of content. E.g. a news article listing page.

Excludes is also now a list, so you can exclude multiple urls.

The image download handling prevents downloading images even when they are the target of a link (not just an image).