Allow multiple matching content type url patterns
The use case is a large site that has some pages which are lists of articles.
Without knowing the list of pages in advance, we should be able to put 2 rules
like:

* | article-summary | ul.list-content li
* | page | .main-content

That will apply in order. If there are no matches for the "ul.list-content li" search,
then the second rule will be tried to extract the main content from the page.