Usability Improvements (#2)
* Add a JSON Viewer for the raw JSON

Add a "copy to clipboard" button for the JSON.

Make the sections of the page expandable and collapsible.

Fix the resolving of relative links in a page.

* Add an meaningful prefix to each client key.

* Use tabs instead of collapsible list items.

* Better display of images

The list of images are not shown as cards, the cors proxy is used if the images are external, and a limit is used so we don't load too many external images in a single page.

* Push the images into the pages array.

This shows which page the images came from.

* Add pagination to all the overview pages.

This means we can get rid of the "max images" check because we are only going to load 10 images at once from the origin site.

* Generate real CLI usage help.

* Force re-building the tabs for a different site.

When the site is changed, the arrays like forms, images etc
are updated. This should update the view, but it is only dependant
on the page number.

Forcing a page number update will reset all tabs to page 0 and
rebuild each tab.

* Allow crawled websites to execute javascript

This allows crawling sites where the content is loaded asynchronously.

This uses jsdom to provide a "virtual browser" for the javascript to use.
It also monkey patches the cors-proxy into the loader so a page can load
it's own external javascript file and run that.

* Fix the duplicate detection in page titles.

Because it is not sorting the array properly, this was leading to random
bits of the titles being removed.

* Use the tag/text ratio as a way to score crawl.

This is done for each page of the site. The final
score is an average of the score for each line of text.

* Save and load the settings for each site.

After viewing a site, use the site settings as
the defaults for the "Add new website" form.

* Apply some filtering based on the score of lines.

* Add a copy button for each form.

This will copy the form text to your clipboard which is easier than a series of text areas in the page.

* Format the JSON nicely.

And remove the logo from the login page.

* Allow multiple matching content type url patterns

The use case is a large site that has some pages which are lists of articles.
Without knowing the list of pages in advance, we should be able to put 2 rules
like:

* | article-summary | ul.list-content li
* | page | .main-content

That will apply in order. If there are no matches for the "ul.list-content li" search,
then the second rule will be tried to extract the main content from the page.

* Add the documents reference to the pages.

Also do not actually attempt to download documents since the downloaded file is not used.

* Change the format of the json.

The document and image elements need to have an "id" element.
It is exactly the same as the url.

* Improvements to score filtering.

Align the order for the settings with the display order
in the form. Only filter non-semantic tags to improve the score.

* Include 'alias' in the json files for each page.

* Improve the structure of the pages.

Include alias and parent attributes for each page, so that the pages
can be constructed into a navigation tree. A menu can be built from this tree.

More settings added for search and replacing text in the source and additional
cleaning.

* Convert absolute URLs to relative.

When we are converting the site to a new domain, we want the URLs
to point to the new domain.

* Allow clients to generate their own login.

Send an email to confirm the account and record the contact information.

* Chow the company name after login.

* Improve the logic for downloading images

Also properly generate the parent pages for navigation

* Do not import images with extremely long URLs.

If the image url is too long it can't be imported.

* All pages must have a title