Large site validation (#3)
* Better structure for the Vue module

Separate the Vue module definition from the HTML source document.

* Adding support for custom fields in content types.

* Updates for GovCMS

- Detect and shorten long URLS
- Map to Gov CMS content types
- Don't clean text until the last step
- Detect Duplicate URLS with different URLS and add a redirect

* Reload some settings that were not in sync.

Also don't skip the first page of a crawl

(because the title matches itself)

* Ignore changes to shortened URLs.

* Support for paginated lists from one site.

The paginated lists from this client site required custom code to step
through the list. There is no "URL" for page 2 etc.

New settings
  - limit the number of URLS in a crawl (useful for testing)
  - formatter for date fields
  - Improvements to the URL shortener (hash the urls in JS, use a flock
    so we don't delete urls from async requests
  - Make the "delay" setting work properly
  - Special handling to extract date fields and translate the date format

* Move the pagination to a custom setting

We can now run scripts on each downloaded page, the pagination example
adds a new URL with a form attached to step through all pages
from a list.

* Improvements to URL shortening.

Shorten the urls in the original text as well. The "words" are so long
the search indexer complains and they need to point to the new site.

* Allow multiple search and replace patterns from a single field.

The search and replace is applied to content as well as urls, so you can map an
entire branch of pages to a different location on the new site.

* Update the gitignore paths.

* Use a curved scale for percentage complete.

* Add a button to download a crawl settings.

Allows you to save the settings without completing a crawl.

* Fix the search replace so it replaces all.

Not just the first one.

* Dont include the content type in the title.

* Allow the site name to be excluded from titles.

The title is for an external page is often the site title and the page name concatenated together.

This not what we want.

* Increase the size of page aliases.

* Test if a url includes a document or image.

Matches more urls than endsWith. Also allow process function to
detect custom forms.

* Better copy feature and empty local storage check

On really large crawls, the copy function struggled with the size and
when localstorage was empty, vue failed to render the page.

* Remove some content cleaning.

On long source documents there was some small content changes which left incomplete documents.
This was fixed by firstly improving the promise variables and secondly by removing the cleaning
from the scoring functions.

There is a new unit test for this.

* Allow a long list of url exclusions